Do not allow duplication of money.
A naïve person might think that one could prevent the duplication of
money by simply not duplicating it.
There is after all nowhere in the code that actually facilitates storing
and restoring the amount of money in the moneypot, especially not in the
duplication handlers.
Therefore it is clearly safe not to do anything about duplicating money,
because it is not duplicated.

The problem with that is that other people try to be helpful.

The NetworkVar system does many useful things to let you develop
entities faster, including creating getters and setters automatically.
Unfortunately it also adds data to the duplicator system.

While this is clearly well meaning and I'm sure very helpful in many
normal cases, this isn't a normal case.

These changes should hopefully prevent any future misunderstandings of
this nature in the future, both by zapping the stored value on create &
restore and also by just blindly setting the money to 0 in every
relevant callback available.

Fun times.