Convert to NetworkVars
The duplicator library was having issues with the stored vars as they were,
so rather than trying to fix it I figured the sensible thing to do would
be to switch to The Right Way To Do It.

Unfortunately since I really don't want to break compatibility with a tool
that was 100% designed to integrate with Luas, I had to add an ugly set
of functions to re-write the `k_*` vars after they're set with Set*.

It works though, so woo.

This does also mean that we can probably start working with the Editable
Entity system which would be neat.