Make the spawn delay timer hang on the last NPC's death
Previously if you had a 10s spawn delay and a max limit of 1, if the NPC
was alive for more than 10s it's replacement would spawn the instant it
died.

This changes things so that the replacement NPC will wait those 10
seconds before spawning.

If anyone preferred it the old way, that is now available with an option
in the "other" field of the toolgun.

(I'm vaguely ashamed of how this patch works, but it's nice and quick)