Add automatic addon versioning
So, the biggest problem I've had with this base_lexentity thing so far
is the really annoying issue of updates.
Garry's Mod's virtual filesystem mechanics mean that if addon A and
addon B both have lua/entities/C, addon A's copy will always override
addon B's copy beacuse it's lexographically smaller.

This means that if I release a new version of addon B that makes use of
a new feature in the base, addon A will have to be updated _first_ in
case a user has both of them installed.

I don't how how good the workshop's delivery system is at making sure
that happens.

I came up with 3 possible methods of dealing with updates to the base:
1) Use a script to release every single addon I publish in advance every
   time I update the base, and then wait a day before publishing
   whatever changes actually needed that update
2) Throw the base away entirely and just go back to copy/pasting
3) Version the base so every addon always has exactly the version that
   it wants to have, even if that means a user ends up with multiple
   copies of the base being used simultaneously

This commit implements #3, since 1 and 2 didn't seem like good ideas.