splat: Refactor the spawn function considerably
I'm sure I'll hate this when I come back to it in another two years, but
the previous version was fairly hard to read due to being absolutely
massive.

I hope this new version with lots of smaller functions called in order
will be a lot easier to understand/update in the future.
Also, I'd like to cal some of these functions from elsewhere, so this
will avoid repeating code 🙂