Add a kill value for NPCs
Moonlight Gaming asked me to make it so they could configure how much
money and XP players got when they killed a NPC spawned by a platform.

The idea was to be able to make boss platforms with high health and
weapon skill and reward players fittingly.

Currently this only works with Vrondakis' addon, but I plan to make a PR
to add this to DarkRP.
Additionally any custom XP/RP systems can use the `KillValue` proprety
to support this.