Add fast entry mode
This allows the keypad's owner to trigger the keypad by looking at it
and hitting enter on their numpad

I 'borrowed' this idea from the unisec addon, though currently it's
fairly unhelpful. I plan to improve on it soon (players, steam friends,
teams, etc etc)

This change also allows me to add a feature later on that will
effectively transform the keypad into a selective entry button that can
also be hacked.