Fix incorrect owner behaviour on spawn platforms
This is a fun one because there were multiple problems simultaneously
causing the same problem.

I've fixed them all in this one commit because I'm not entirely sure
which one was entirely to blame and I can't be bothered to split the
changes.

This is technically also breaking changes but I'm being cavaleer today